family-studies - Improve support for family studies Moved JSONResponse and StatusResponse to org.phenotips.studies.family.script.
